+++ workspace/CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+* `workspace.list`: support filtering by dependency
 * `workspace.list`: add `--maintainer` for filtering projects with the given maintainer
 
 ## [v0.1.1](https://github.com/sportradar/elixir-workspace/tree/workspace/v0.1.1) (2024-07-04)